0518 api 관련 기능을 kodeal 디렉터리로 옮김
