+++ kodeal/views.py
@@ -1,9 +1,9 @@
 import ssl
 import urllib
 import urllib.request
+import json
 
 
-from blog.views import extract_question_sentences
 from config.my_settings import CLIENT_ID, CLIENT_SECRET
 
 # Create your views here.
@@ -31,3 +31,49 @@ def papago(text):
     else:
         print("Error Code:" + rescode)
         return 'ERROR'
+
+# Codex의 입력으로 넣을 문장을 번역한 결과(Papago) 중에서 사용자의 질문에 해당하는 부분만 추출하는 함수
+def extract_question_sentences(response):
+    json_object = json.loads(response)
+    return json_object['message']['result']['translatedText']
+
+
+# Codex로부터 반환된 answer 값 전체 중에서 사용자의 질문에 대한 답변만 추출하는 함수
+def extract_answer_sentences(response):
+    # 반환된 response 중에서 질문에 대한 답변이 포함된 'choices' 부분만 get
+    choices = json.dumps(*response['choices'])
+
+    # 위의 과정에서 choices의 값은 str 타입이기 때문에 JSON 형태로 변환해야 함
+    json_choices = json.loads(choices)
+
+    # JSON 형태로 변환된 문자열 중 키가 'text'인 값을 return
+    answer = json_choices['text']
+
+    # 전처리 과정을 거친 결과 반환
+    return perform_preprocessing(answer)
+
+
+# ※※※ 전처리 기능을 총괄하는 함수 ※※※
+def perform_preprocessing(answer):
+    # 문장 앞뒤로 불필요한 문자 제거
+    answer = remove_unnecessary_char(answer)
+
+    return answer
+
+
+# answer 문장 앞뒤로 불필요한 문자 제거
+def remove_unnecessary_char(sentence):
+    # 첫 글자가 콜론(:)이라면 제거
+    def remove_first_colon(answer):
+        if answer[0] == ':':
+            return answer[2:]
+        else:
+            return answer
+
+    # 결과로 전달되는 answer 문장에서 맨 앞의 개행 문자 전처리
+    def remove_two_newline_char(answer):
+        return answer.strip()
+
+    sentence = remove_first_colon(sentence)
+    sentence = remove_two_newline_char(sentence)
+    return sentence
